Adding the Pagination and sorting for Products, also some refactoring for packages
