+++ src/main/java/com/codegym/cgzgearservice/dto/ProductDTO.java
@@ -1,10 +1,8 @@
 package com.codegym.cgzgearservice.dto;
 
-import com.codegym.cgzgearservice.model.entitiy.product.ProductDetail;
 import lombok.Data;
 
 import java.util.List;
-import java.util.Set;
 
 @Data
 public class ProductDTO {
