+++ src/main/java/com/codegym/cgzgearservice/entitiy/product/Cartline.java
@@ -1,0 +1,31 @@
+package com.codegym.cgzgearservice.entitiy.product;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "cartlines")
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class Cartline {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "product_id")
+    private Product product;
+
+    @Column(name = "quantity")
+    private Integer quantity;
+
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "cart_id")
+    private Cart cart;
+}
