+++ src/main/java/com/codegym/cgzgearservice/entitiy/product/Order.java
@@ -1,6 +1,5 @@
-package com.codegym.cgzgearservice.model.entitiy.product;
+package com.codegym.cgzgearservice.entitiy.product;
 
-import com.codegym.cgzgearservice.model.entitiy.product.Cart;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
