--- src/main/java/com/codegym/cgzgearservice/model/entitiy/product/SpecificationTemplate.java
@@ -1,28 +1,0 @@
-package com.codegym.cgzgearservice.model.entitiy.product;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "specification_templates")
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class SpecificationTemplate {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "category_id")
-    private Category category;
-
-    @Column(name= "spec_key")
-    private String specKey;
-
-}
