+++ src/main/java/com/codegym/cgzgearservice/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.codegym.cgzgearservice.repository;
 
-import com.codegym.cgzgearservice.model.entitiy.product.Category;
+import com.codegym.cgzgearservice.entitiy.product.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category,Long> {
