+++ src/main/java/com/codegym/cgzgearservice/repository/ProductDetailRepository.java
@@ -1,13 +1,7 @@
 package com.codegym.cgzgearservice.repository;
 
-import com.codegym.cgzgearservice.model.entitiy.product.Product;
-import com.codegym.cgzgearservice.model.entitiy.product.ProductDetail;
+import com.codegym.cgzgearservice.entitiy.product.ProductDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
 public interface ProductDetailRepository extends JpaRepository<ProductDetail,Long> {
     ProductDetail findProductDetailByProductId(Long productId);
 }
