+++ src/main/java/com/codegym/cgzgearservice/repository/ProductRepository.java
@@ -1,13 +1,9 @@
 package com.codegym.cgzgearservice.repository;
 
-import com.codegym.cgzgearservice.model.entitiy.product.Product;
+import com.codegym.cgzgearservice.entitiy.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-
 import java.util.List;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
-
+//    List<Product> findAllByCategory(String category);
 }
