+++ src/main/java/com/codegym/cgzgearservice/repository/RoleRepository.java
@@ -1,10 +1,7 @@
 package com.codegym.cgzgearservice.repository;
 
-import com.codegym.cgzgearservice.model.entitiy.user.Role;
+import com.codegym.cgzgearservice.entitiy.user.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
 public interface RoleRepository extends JpaRepository<Role,Long> {
     Role findRoleByName( String name);
 }
