+++ src/main/java/com/codegym/cgzgearservice/repository/SpecificationTemplateRepository.java
@@ -1,6 +1,6 @@
 package com.codegym.cgzgearservice.repository;
 
-import com.codegym.cgzgearservice.model.entitiy.product.SpecificationTemplate;
+import com.codegym.cgzgearservice.entitiy.product.SpecificationTemplate;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
