+++ src/main/java/com/codegym/cgzgearservice/service/ProductService.java
@@ -1,7 +1,8 @@
 package com.codegym.cgzgearservice.service;
 
 import com.codegym.cgzgearservice.dto.ProductDTO;
-import com.codegym.cgzgearservice.model.entitiy.product.Product;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
@@ -9,6 +10,6 @@ public interface ProductService {
     ProductDTO createProduct(ProductDTO productDTO);
     ProductDTO updateProduct(Long productId, ProductDTO productDTO);
     ProductDTO getProductById(Long productId);
-    List<ProductDTO> getAllProducts();
+    Page<ProductDTO> getAllProducts(Pageable pageable);
     ProductDTO deleteProduct(Long productId);
 }
