+++ src/main/java/com/codegym/cgzgearservice/service/UserService.java
@@ -2,7 +2,7 @@ package com.codegym.cgzgearservice.service;
 
 
 import com.codegym.cgzgearservice.dto.UserDTO;
-import com.codegym.cgzgearservice.model.entitiy.user.User;
+import com.codegym.cgzgearservice.entitiy.user.User;
 
 import java.util.List;
