+++ src/main/java/com/codegym/cgzgearservice/service/impl/SecurityServiceImpl.java
@@ -1,4 +1,5 @@
 package com.codegym.cgzgearservice.service.impl;
 
-public class SercurityServiceImpl {
+public class SecurityServiceImpl {
+
 }
