+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.flyway.url=jdbc:mysql://localhost:3306/zgear
 spring.flyway.user=root
 spring.flyway.password=12345678
-spring.flyway.enabled=true
+spring.flyway.enabled=false
 
 spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
